JPERF-729 Move CrowdEncryptedPasswordProvider to a separate file
